pyproject.toml
[project]
name = "rag-indepth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
]
